Support opening multiple files simultaneously.
Add support for loading just the file headers and not the entire elements data.
Add support for multiple simultaneous DLG-3 files loaded (drawn side-by-side).
Pass a mapname as argument to launch mapv with half a quadrangle loaded.
Remove the attempt at doing MVC, not worth the trouble here because the model
never changes independently of the view.